# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.UserRole": {
				"type": "string",
				"enum": [
					"ADMIN",
					"CASHIER",
					"STAFF"
				]
			},
			"UserRole": {
				"$ref": "#/components/schemas/_36_Enums.UserRole"
			},
...
			"_36_Enums.OrderStatus": {
				"type": "string",
				"enum": [
					"PENDING",
					"PREPARING",
					"SERVED",
					"PAID",
					"CANCELLED"
				]
			},
			"OrderStatus": {
				"$ref": "#/components/schemas/_36_Enums.OrderStatus"
			},
			"OrderItemResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"menuItemId": {
						"type": "string"
					},
					"menuItemName": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"menuItemId",
					"quantity",
					"unitPrice",
					"note"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"orderNumber": {
						"type": "number",
						"format": "double"
					},
					"tableId": {
						"type": "string"
					},
					"tableNumber": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/OrderStatus"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string",
						"nullable": true
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderItemResponseDto"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"orderNumber",
					"tableId",
					"userId",
					"status",
					"totalAmount",
					"note",
					"items",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateOrderItemDto": {
				"properties": {
					"menuItemId": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"note": {
						"type": "string"
					}
				},
				"required": [
					"menuItemId",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateOrderDto": {
				"properties": {
					"tableId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/CreateOrderItemDto"
						},
						"type": "array"
					}
				},
				"required": [
					"tableId",
					"userId",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateOrderStatusDto": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/OrderStatus"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MenuItemResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string",
						"nullable": true
					},
					"isAvailable": {
						"type": "boolean"
					},
					"categoryId": {
						"type": "string"
					},
					"categoryName": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"price",
					"imageUrl",
					"isAvailable",
					"categoryId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMenuItemDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"isAvailable": {
						"type": "boolean"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"price",
					"categoryId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMenuItemDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"isAvailable": {
						"type": "boolean"
					},
					"categoryId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CategoryResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"isAvailable": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"isAvailable",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isAvailable": {
						"type": "boolean"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"isAvailable": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponseDto": {
				"properties": {
					"token": {
						"type": "string"
					},
					"user": {
						"properties": {
							"role": {
								"$ref": "#/components/schemas/UserRole"
							},
							"fullName": {
								"type": "string"
							},
							"email": {
								"type": "string",
								"nullable": true
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"role",
							"fullName",
							"email",
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"token",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"username",
					"password",
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginDto": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PinLoginDto": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"pin": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"pin"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/orders": {
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderDto"
							}
						}
					}
				}
			}
		},
		"/api/orders/{id}": {
			"get": {
				"operationId": "GetOrderById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Order not found"
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/orders/{id}/status": {
			"put": {
				"operationId": "UpdateOrderStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Order not found"
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOrderStatusDto"
							}
						}
					}
				}
			}
		},
		"/api/orders/items/{itemId}/status": {
			"patch": {
				"operationId": "UpdateOrderItemStatus",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Order"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"type": "string"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/menu-items": {
			"get": {
				"operationId": "GetMenuItems",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MenuItemResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MenuItem"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMenuItem",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"MenuItem"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMenuItemDto"
							}
						}
					}
				}
			}
		},
		"/api/menu-items/{id}": {
			"get": {
				"operationId": "GetMenuItemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "MenuItem not found"
					}
				},
				"tags": [
					"MenuItem"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMenuItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MenuItemResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "MenuItem not found"
					}
				},
				"tags": [
					"MenuItem"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMenuItemDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMenuItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"MenuItem"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto"
							}
						}
					}
				}
			}
		},
		"/api/categories/{id}": {
			"get": {
				"operationId": "GetCategoryById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Category not found"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CategoryResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Category not found"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Category"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Registered Successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/pin-login": {
			"post": {
				"operationId": "PinLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PinLoginDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/me": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}